Add support to serialize document to oscal
This commit adds support to transformed the `Coradoc::Document` to
`Oscal` document. As an initial step, it builds a hash structure that
was proposed by `Oscal`, and then use the default `yaml` library to
convert that to yaml.

This covers the top level `metadata` and `groups` attributes, along
with controls in the document. For each control, it also goes into
the clauses and then list out the contents or different parts.

The parsing is still missing the followings:

- List items in the content blocks
- Table and their rows and columns
- Nested structure for each sub-content