Add support to parse block types
This commit adds support to parse different types of blocks. So,
now whenever there is a block in the content, then it should parse
that as expected.

To simplify the process, this commit also extract the block style
in a way then we can easily extend and test new block types.