[WIP] [DO-NOT-MERGE] Toolset for combining coradoc & reverse_adoc
This PR is NOT intended to be merged. It is merely a showcase of
our current work at merging reverse_adoc into coradoc. A request
for comments of sorts.

We will proceed with this merger once all remaining pull requests
in reverse_adoc repository have been merged.

The approach is to use a migration script to:
1. Modify reverse_adoc, to create a repository in a state of no
   files having a duplicate name with coradoc repo
2. Rename ReverseAdoc class to Coradoc::ReverseAdoc
3. Git merge reverse_adoc into coradoc
4. Merge other contents manually
5. Finally, change reverse_adoc repository into a stub repo, so
   that users of reverse_adoc will have a clean path to update.

This approach will preserve git history for both repositories,
merging them together.