Don't add a space before link if constrained.
By this, we mean - if before a link there's a space, or beginning
of a block, we don't need to add another space.

In fact, we shouldn't, because in a case of code like...

<div><a href="test">test</a></div>

If we add a space before a link, we open a code block and thus
we just get a source code and not a link.