Refactor config, add time tracking facility, optimize cleaner
In particular, I was curious what caused a performance problem on
a large document I'm working on. Turned out, it was a
remove_inner_whitespace procedure in Cleaner. With a simple fix I
managed to make it finish in 1 second, instead of 170s.

All the rest of the processing combined takes 10s, so we will be
able to progress much faster on next issues.