Simplify table cells that contain DIVs
The idea here, is that HTML content generators may often introduce
a lot of unnecessary markup, that only makes sense in the HTML+CSS
context. The idea is that certain cases can be simplified, making it
so that the result is equivalent, but much simpler, allowing us to
generate a nicer AsciiDoc syntax for those cases.