Change logic of generating one-line table rows
Let's move the logic of delimiting tables to Coradoc, as I think it
makes more sense to be there. This changes semantics a little - now
one-line rows are generated if there are any AsciiDoc cells. Before
that, it was a logic of Cell to decide if it wanted to be generated
multiline or not. This results in nicer tables.