Ensure row/column table integrity
We can't be sure that a table from a document is correct - and most
often we stumble upon those that aren't. Either colspans or rowspans
are often wrong, in which case we need to shorten them. Sometimes
not enough cells are provided. Then we need to add virtual cells.

This is because in AsciiDoc, in certain cases, we provide cells
sequentially, without real diversification of rows and we need
to provide a number of cells upfront. This commit introduces a
quite advanced algorithm to correct the tables while they are still
in HTML and also robustly computes a number of cells.